refactor delete collection to collection actions component
